refactor: Add email to email_list table in supabaseClientFunctions.ts
Add the functionality to add an email to the email_list table in the supabaseClientFunctions.ts file. This allows for the insertion of a new email into the database. The code has been added to the existing supabaseClientFunctions.ts file, and an error handling mechanism has been implemented. This change improves the functionality of the application by enabling the addition of emails to the email list.